app/requirements.txt
websockets>=10.4
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
numpy>=1.21.0
redis>=4.3.4
requests>=2.26.0
python-dotenv>=0.19.0
asyncio>=3.4.3
PyYAML>=6.0
